Added photos functionality

+++ AutoDealer/AutoDealer.Miscellaneous/Interfaces/IFileAttachment.cs
@@ -6,7 +6,7 @@ namespace AutoDealer.Miscellaneous.Interfaces
     public interface IFileAttachment
     {
         string FileName { get; }
-        int FileSize { get; }
+        long FileSize { get; }
         Task CopyToAsync(FileStream stream);
     }
 }
